Moved common server functions out of flask
